Add error handling to API lookups, OpenAlex returns
Incorporated more "try, except" calls for the info from API lookups in case the DOI doesn't exist in the database or lookup returns an error. Additionally began to setup using OpenAlex as a tool for filling in information that is returned as errors from Crossref/Elsevier/Semantic Scholar. Not quite complete but a start if we want to expand on it.